fixed test failure, renamed top level directories
