+++ proteuslib/flowsheets/full_treatment_train/analysis/flowsheet_single_stage.py
@@ -22,13 +22,13 @@ from idaes.core import FlowsheetBlock
 from idaes.core.util.initialization import propagate_state
 from idaes.core.util.scaling import calculate_scaling_factors
 
-from proteuslib.flowsheets.full_treatment_train.example_flowsheets import (desalination,
-                                                                           feed_block,
-                                                                           gypsum_saturation_index,
-                                                                           translator_block,
-                                                                           costing,
-                                                                           financials,)
-from proteuslib.flowsheets.full_treatment_train.example_models import property_models
+from proteuslib.flowsheets.full_treatment_train.flowsheet_components import (desalination,
+                                                                             feed_block,
+                                                                             gypsum_saturation_index,
+                                                                             translator_block,
+                                                                             costing,
+                                                                             financials, )
+from proteuslib.flowsheets.full_treatment_train.model_components import property_models
 from proteuslib.flowsheets.full_treatment_train.util import (solve_with_user_scaling,
                                                              check_dof)
