+++ proteuslib/flowsheets/full_treatment_train/chemistry_flowsheets/posttreatment_ideal_naocl_chlorination_block.py
@@ -70,7 +70,7 @@ from pyomo.util.check_units import assert_units_consistent
 
 
 from proteuslib.flowsheets.full_treatment_train.util import solve_with_user_scaling, check_dof
-from proteuslib.flowsheets.full_treatment_train.example_models import property_models
+from proteuslib.flowsheets.full_treatment_train.model_components import property_models
 from idaes.core.util import get_solver
 
 # Import the idaes objects for Generic Properties and Reactions
@@ -83,10 +83,10 @@ from idaes.generic_models.properties.core.generic.generic_reaction import (
 from idaes.generic_models.unit_models.equilibrium_reactor import EquilibriumReactor
 
 # Import the Proteuslib object inherited for the Mixer unit model
-from proteuslib.flowsheets.full_treatment_train.example_models import Mixer
+from proteuslib.flowsheets.full_treatment_train.model_components import Mixer
 
 # Import costing and financials to test
-from proteuslib.flowsheets.full_treatment_train.example_flowsheets import costing, financials
+from proteuslib.flowsheets.full_treatment_train.flowsheet_components import costing, financials
 
 from idaes.generic_models.unit_models.translator import Translator
