+++ proteuslib/flowsheets/full_treatment_train/flowsheet_components/flowsheet_limited_softening.py
@@ -24,12 +24,12 @@ from idaes.core.util.scaling import (calculate_scaling_factors,
                                      badly_scaled_var_generator,
                                      constraint_autoscale_large_jac)
 from idaes.core.util.initialization import propagate_state
-from proteuslib.flowsheets.full_treatment_train.example_flowsheets import (pretreatment_softening,
-                                                                           desalination,
-                                                                           translator_block,
-                                                                           costing,
-                                                                           financials)
-from proteuslib.flowsheets.full_treatment_train.example_models import property_models
+from proteuslib.flowsheets.full_treatment_train.flowsheet_components import (pretreatment_softening,
+                                                                             desalination,
+                                                                             translator_block,
+                                                                             costing,
+                                                                             financials)
+from proteuslib.flowsheets.full_treatment_train.model_components import property_models
 from proteuslib.flowsheets.full_treatment_train.util import solve_with_user_scaling, check_dof
 
 # Added import statements for testing.
