+++ proteuslib/flowsheets/full_treatment_train/flowsheet_components/sweep_flowsheet_limited.py
@@ -4,7 +4,7 @@ from mpi4py import MPI
 from idaes.core.util import get_solver
 
 from proteuslib.tools.parameter_sweep import LinearSample, parameter_sweep
-from proteuslib.flowsheets.full_treatment_train.example_flowsheets.flowsheet_limited import (solve_optimization, optimize, set_up_optimization)
+from proteuslib.flowsheets.full_treatment_train.flowsheet_components.flowsheet_limited import (solve_optimization, optimize, set_up_optimization)
 
 comm = MPI.COMM_WORLD
 rank = comm.Get_rank()
