+++ proteuslib/flowsheets/full_treatment_train/flowsheet_components/sweep_flowsheet_mvp.py
@@ -4,7 +4,7 @@ from scipy.interpolate import griddata
 from idaes.core.util import get_solver
 
 from proteuslib.tools.parameter_sweep import _init_mpi, LinearSample, parameter_sweep
-from proteuslib.flowsheets.full_treatment_train.example_flowsheets.flowsheet_mvp import (solve_optimization, optimize, set_up_optimization)
+from proteuslib.flowsheets.full_treatment_train.flowsheet_components.flowsheet_mvp import (solve_optimization, optimize, set_up_optimization)
 
 comm, rank, num_procs = _init_mpi()
