+++ proteuslib/flowsheets/full_treatment_train/flowsheet_components/tests/test_pretreatment_softening.py
@@ -13,8 +13,8 @@
 import pytest
 from pyomo.environ import ConcreteModel, value, TransformationFactory
 from idaes.core import FlowsheetBlock
-from proteuslib.flowsheets.full_treatment_train.example_flowsheets import pretreatment_softening
-from proteuslib.flowsheets.full_treatment_train.example_models import property_models
+from proteuslib.flowsheets.full_treatment_train.flowsheet_components import pretreatment_softening
+from proteuslib.flowsheets.full_treatment_train.model_components import property_models
 from proteuslib.flowsheets.full_treatment_train.util import check_build, check_dof, check_scaling
