+++ proteuslib/flowsheets/full_treatment_train/flowsheet_components/tests/test_translator_block.py
@@ -13,8 +13,8 @@
 import pytest
 from pyomo.environ import ConcreteModel
 from idaes.core import FlowsheetBlock
-from proteuslib.flowsheets.full_treatment_train.example_flowsheets import translator_block
-from proteuslib.flowsheets.full_treatment_train.example_models import property_models
+from proteuslib.flowsheets.full_treatment_train.flowsheet_components import translator_block
+from proteuslib.flowsheets.full_treatment_train.model_components import property_models
 from proteuslib.flowsheets.full_treatment_train.util import check_scaling
