+++ proteuslib/flowsheets/full_treatment_train/model_components/eNRTL/test_enrtl.py
@@ -16,7 +16,7 @@ from idaes.core import FlowsheetBlock
 from idaes.generic_models.properties.core.generic.generic_property import GenericParameterBlock
 from idaes.core.util.scaling import calculate_scaling_factors, get_scaling_factor, constraint_scaling_transform
 from idaes.core.util import get_solver
-from proteuslib.flowsheets.full_treatment_train.example_models.eNRTL import entrl_config_FTPx, entrl_config_FpcTP
+from proteuslib.flowsheets.full_treatment_train.model_components.eNRTL import entrl_config_FTPx, entrl_config_FpcTP
 from proteuslib.flowsheets.full_treatment_train.util import check_scaling, solve_with_user_scaling
 
 def simulate_enrtl_FTPx(state_var_args):
