+++ proteuslib/flowsheets/full_treatment_train/model_components/tests/test_prop_pack.py
@@ -19,8 +19,8 @@ from idaes.core import FlowsheetBlock
 import idaes.core.util.scaling as iscale
 from pyomo.util.check_units import assert_units_consistent
 from proteuslib.flowsheets.full_treatment_train.util import solve_with_user_scaling, check_dof
-import proteuslib.flowsheets.full_treatment_train.example_models.seawater_ion_prop_pack as property_seawater_ions
-import proteuslib.flowsheets.full_treatment_train.example_models.seawater_salt_prop_pack as property_seawater_salts
+import proteuslib.flowsheets.full_treatment_train.model_components.seawater_ion_prop_pack as property_seawater_ions
+import proteuslib.flowsheets.full_treatment_train.model_components.seawater_salt_prop_pack as property_seawater_salts
 
 # -----------------------------------------------------------------------------
 @pytest.mark.component
