+++ proteuslib/flowsheets/full_treatment_train/model_components/unit_1DRO.py
@@ -16,7 +16,7 @@
 from pyomo.environ import ConcreteModel
 from idaes.core import FlowsheetBlock
 from idaes.core.util.scaling import calculate_scaling_factors
-from proteuslib.flowsheets.full_treatment_train.example_models import property_models
+from proteuslib.flowsheets.full_treatment_train.model_components import property_models
 from proteuslib.unit_models.reverse_osmosis_1D import (ReverseOsmosis1D,
                                                        ConcentrationPolarizationType,
                                                        MassTransferCoefficient,
